feat: Major refactor for performance improvements
