+++ src/app/[locale]/impressum/page.tsx
@@ -1,29 +1,28 @@
-"use client";
-
-import { useTranslations } from "next-intl";
-import { useState, useEffect } from "react";
+import { getTranslations } from "next-intl/server";
 
 import Container from "@/components/elements/container";
-import Nav from "@/components/nav";
 
-export default function Impressum() {
-  const t = useTranslations();
-  const [impressum, setImpressum] = useState("");
+async function getImpressum() {
+  try {
+    const response = await fetch("https://philipbrembeck.com/impressum.txt");
+    if (!response.ok) {
+      throw new Error(`Failed to fetch impressum: ${response.status}`);
+    }
+    return response.text();
+  } catch (error) {
+    console.error("Failed to fetch impressum:", error);
+    return "";
+  }
+}
 
-  useEffect(() => {
-    fetch("https://philipbrembeck.com/impressum.txt")
-      .then((response) => response.text())
-      .then((text) => setImpressum(text))
-      .catch((error) => console.error(error));
-  }, []);
+export default async function Impressum() {
+  const t = await getTranslations();
+  const impressum = await getImpressum();
 
   return (
-    <>
-      <Nav />
-      <Container heading={t("More.imprint")}>
-        <p className="small">{t("Privacy.germanonly")}</p>
-        <div dangerouslySetInnerHTML={{ __html: impressum }} />
-      </Container>
-    </>
+    <Container heading={t("More.imprint")}>
+      <p className="small">{t("Privacy.germanonly")}</p>
+      <div dangerouslySetInnerHTML={{ __html: impressum }} />
+    </Container>
   );
 }
