+++ src/app/[locale]/ingredients/page.tsx
@@ -1,14 +1,11 @@
 import Container from "@/components/elements/container";
-import IngredientsCheck from "@/components/ingredientscheck";
-import Nav from "@/components/nav";
+import { IngredientsForm } from "@/components/IngredientsCheck/IngredientsForm";
 
 export default function IngredientsPage() {
   return (
     <>
-      <div id="modal-root"></div>
-      <Nav />
       <Container logo={false} backButton={false}>
-        <IngredientsCheck />
+        <IngredientsForm />
       </Container>
     </>
   );
