+++ src/components/Check/models/product.ts
@@ -1,0 +1,13 @@
+export interface NutriscoreGrade {
+  score: string;
+  className: string;
+}
+
+export interface ProductState {
+  vegan: string;
+  vegetarian: string;
+  animaltestfree: string;
+  palmoil: string;
+  nutriscore: NutriscoreGrade;
+  grade: NutriscoreGrade;
+}
