+++ src/components/IngredientsCheck/models/IngredientResult.ts
@@ -1,0 +1,6 @@
+export interface IngredientResult {
+  vegan: boolean | null;
+  surelyVegan: string[];
+  notVegan: string[];
+  maybeVegan: string[];
+}
