+++ src/components/Scanner/index.ts
@@ -1,0 +1,1 @@
+export { ScanButton as default } from "./ScanButton";
