+++ src/components/Scanner/models/scanner.ts
@@ -1,0 +1,10 @@
+export interface ScannerProps {
+  onDetected: (result: DetectionResult) => void;
+  setScanning: (scanning: boolean) => void;
+}
+
+export interface DetectionResult {
+  codeResult: {
+    code: string;
+  };
+}
